Added .gitignore files
Makes git ignore build files and preserve directories that the build
process expects to be empty.  The Makefiles should probably be modified
in the future to not require such directories to be pre-existing in the
first place, but for now this is a work-around of git's (intentional)
inability to store empty directories.