Documentation: rewrote most of README, reformatted COPYING
COPYING is reformatted to fill in 72-character lines efficiently, and
the copyright dates are expanded which has some legal benefits.  None of
the license text itself is actually changed.

README rewritten in AsciiDoc format, the original contents are used as
the preamble.  Expanded to explain a bit more about Free Software and
how to contribute levels (still need to write specifics about maps and
other resources, actually submitting work, etc...).