Add copyrights, remove latest/ directory
Adding copyright notices to critical files; some files which the project
is not strictly dependent on, such as tools/cleanroom, are not tagged;
many of these were meant for temporary use in assisting the development
of Freedoom and their existence should probably be questioned.

latest/ directory is removed since it's not critical to building the
project, its use was for pre-sourceforge daily builds.  Such a frequent
building system may still be sought after, but this doesn't need to be
in the repo ;)

Signed-off-by: Mike Swanson <mikeonthecomputer@gmail.com>