map13: fix exiting the level; thanks anonymous bug reporter
This map could not be completed as it had no exits. The latest did have
and exit but could not be reached. A silent teleport line before it.
This has been rectified. A number of other tweaks & fixes to the level
also. Lines 126,119 and 122 had the NUKEMAP incorrectly on the mid tex.

Signed-off-by: Mike Swanson <mikeonthecomputer@gmail.com>