textures: fix wall texs for consistency; thanks anonymous bug reporter
Included is a new set of textures for gstfont1-3 (wall58_[123].gif);
Compare the blood from the mouth in the old set.  This is reflected
(fixed) for secure 432 of map 25.

Also included are new patches wall52_2 and wall54_2 (cement5 and cement6
respectively).  The lower orange and black striping now matches the
other "cement" textures.  Check the start of map 10.

Signed-off-by: Mike Swanson <mikeonthecomputer@gmail.com>