Change example commit in README and fix formatting in asciidoc.
The commit referenced was from prior to the history-rewrite and
no longer was a real commit in the repository.

Also, asciidoc had some formatting error in the "Submitting your
work" section where a large chungy of text was being rendered as
monospace in the HTML version.  To fix this, "git commit -s" is
surrounded by +s rather than `s.

Signed-off-by: Mike Swanson <mikeonthecomputer@gmail.com>