buildcfg: use real sprites in FreeDM iwad
Remove the #ifdef FREEDM section that makes all monsters appear as a
black silhouette. This leaves sprites/{blank,nomonst}.gif unused.

I believe the plan is that FreeDM will become a vanilla-compatible IWAD
that the user can use with Chocolate Doom to play single player vanilla
maps, so real monsters are needed in it.

Signed-off-by: RjY <rjy@users.sourceforge.net>