levels: dm02: stop using GRAY4 as a two-sided middle texture
Using a multi-patch texture on the middle texture of a two-sided linedef
causes the so-called medusa effect in vanilla, so in FreeDM it is a bug.

In dm02, we have several lines which use GRAY4 as a middle texture, with
silver borders above and below. These are all in front of a thin sector
backed by a one-sided line with GRAYBIG on it. So:

- set two-sided lines to have middle texture "-" instead of GRAY4;
- set one-sided lines to have middle texture GRAY4 instead of GRAYBIG;
- adjust a few texture alignments.

The sector formerly behind the GRAY4 lines is now visible as a thin
silver line in front of them, above and below the silver borders.