levels: e1m1: fix texture alignment on exit switch
Swap the offsets of the two segs on the front side of line 2544, the
exit switch. Seg 235 (from 0 to 48) had offset 48, and 236 (48 to 64)
had offset 0. Clearly these are the wrong way round.

Only noticable in engines that do not use the GL nodes in this map,
and have not yet switched to recalculating seg offsets themselves,
after that spate of wads with bad seg offsets circa late 2005.