levels: E1M7: move walk trigger to other side of door (Patrol1985)
A classic ZDoomism (implicit passuse): A door (sector 204) could not be
opened due to its west side (linedef 1208) being blocked by a walk
trigger (linedef 1201). This left both door unable to be opened, and
walk trigger unable to be activated. The latter left the player stuck in
the immediate area, unable to finish the level.

The simplest fix, applied here, is to move the walk trigger to the other
side of the door (linedef 1214). The player must still walk through the
door and "see" the blue key vanish to escape the maze.

Thanks to Patrol1985 (and others) for reporting this:
http://www.doomworld.com/vb/post/1209971