dbimpact: avoid use of tag 666 on M8
There are four switches that all need to be pressed to activate the
finale of the map. In the original these would cause four barons to be
crushed, the final one causing several floors to lower. Since this use
of barons makes the map depend on the E1M8 slot, it needs reworking.

- Change the switches to raise floor to lowest adjacent ceiling types
  (thus creating an instant floor lowering)

- Near to each switch place lines that lower the former tag 666 floors
  to their _highest_ adjacent.

Thus these floors only lower when all four switches have been pressed
and all four control sectors have been lowered.

Obviously this could be done with voodoo doll scripts or using multiple
switches and the Boom passuse flag but since dbimpact was designed for
Doom complevel I did not want to risk making it require Boom complevel.

Make a couple of other adjustments to support this:

- The four switch-pressed indicators on either side of the exit door now
  unfortunately no longer start moving (they're not crushers)

- "Crack" the northwest demon face tower so there is a tiny invisible
  hole between the bright part and the small shadow on the corner.
  Otherwise lower to highest adjacent floor would not work as required.

- Also, make all control sectors invisible on the automap (after all,
  there is a computer map on this level)