levels: E1M5 2013-12-15 update by Z0k
-moved an imp that was stuck on a console panel near of the swicht,
preveting the imp for moving futher.

-Size of the columns that hold tech pillar now are reduced, this for
preveting player getting stuck in middle of them

-Added 2 to 5 % damage to the acid pit of the tech pillars

-Added a new teleporter on the red key zone to reach the swicht that
open the door of the exit, this to help the player who dont know that
the teleporters where monsters come out are teleporters they can use. so
players can use both teleporters, one which is useful as shortcut to end
the level quickly and the other to keep the way of the level delevop

[ RjY: changes to submitted version:
  - remove spurious dehacked lump (cf. commit 618b444049)
  - set passuse flag on lines 1795, 1800, and 1806, so the lift can be
    operated from below, and the door opened from above ]

DW: http://www.doomworld.com/vb/post/1222853
Signed-off-by: RjY <rjy@users.sourceforge.net>