README: Reflect the project's actual methods
When I wrote the README before, I was basically demanding the use of
Git's signed-off messages, which is just wrong-headed. Also change the
example commit so it uses an action stamp, no more hard dependency on
a sha1 identifier that may change if the repository is rewritten (has
happened) or moved to another VCS (has happened, also); additionally,
it's just a better example commit to use. :-)

Remove the "Build system" section which encouraged a "Do Not Touch"
philosophy which is also wrong-headed. Any change for improvement is
self-evident: should not tell people to not bother without explaining
themselves.

Also, rename the file so GitHub renders it nicely.