music: Add music for FreeDM levels.
As part of #10, link in some music tracks to fit FreeDM's levels,
using some of the currently-unused MIDI tracks in the tree. The
choices are made based on my own subjective opinion to try to fit
the levels we have.

DM17 is still missing music, and there is no music for the empty
level slots. Generally, the best music tracks have been used for
the early levels and the later ones are not as good. There's still
a lot of room for improvement.