levels: allow wallrunning and move a teleporter pad
Rotated the level 270° to accomidate for a bug in the Doom engine that
allows superspeeds against some walls, abuse this bug in map01 on the
long edges :)

Also, move a teleporter exit in the Things room onto the pad sector
itself. This makes it so you don't just teleport back by strafing
right immediately.