graphics: remove the date from the titlepic
This is a small step towards deterministic builds (#41). It's not very
useful information on its own: what it tells, you can gather by the
file modification date (which is preserved in the zip files), and even
better (the titlepic date was based on the builder's local time, not
UTC!). Some people have tended to refer to the date as well when
talking about what build they have, which is also unhelpful, multiple
commits can happen per day and the Version marker already uses git
describe for more accurate information.