patches: Replace w103_[12].
I made these textures in the early days of the project and I'm pretty
sure the symbols came from an animation on a website I found. The
symbols are pretty generic alchemical symbols but it still bothers
me on the rare occasion that I see the textures used in a level.
Recreate them using some public domain imagery in their place.