dehacked: Add ZDoom obituaries.
When playing in ZDoom-based ports, a message is shown when the player
dies which depends on the monster that killed them or (in deathmatch)
the weapon they were killed with. Because ZDoom uses obituaries that
mention the Doom weapons and monsters, override so that Freedoom is
internally consistent when playing using these ports.