textgen: Adjust M_MULTI graphic height, alignment.
M_MULTI is used in some multiplayer ports (Zandronum, others?) for
the main menu "Multiplayer" option. Make this taller to match the
other main menu graphics and align appropriately in buildcfg.

This conflicts with the height of the graphic as found in other ports
like SMMU/Eternity/PrBoom, but I've checked, and in those ports it's
either unused or short enough that it still works okay.