Remove doom2tr from CREDITS.
The Doom 2 retexturing project was a project that Freedoom partnered
with many years ago in an attempt to develop high resolution textures
for modern 3D source ports. The idea was that there was an overlap in
goals between Freedoom wanting to create new compatible textures, and
doom2tr wanting to develop high resolution textures.

However, this partnership never really worked out and as far as I
remember nothing was ever contributed to Freedoom from the doom2tr
side. It seems inaccurate to keep it in the CREDITS file; I checked
the source tree and couldn't find any reference to it.