sprites: Adjust offset for fist sprites.
Freedoom's fist punches from the right side of the screen; however,
some source ports can run in widescreen modes that can cause the
punch animation frames to be cut off at the right side of the screen.
Move all fist sprites 50 pixels to the right so that this doesn't
happen.

I tested this in ZDoom running in a 16:9 screen aspect ratio and I
was unable to cause any noticeable cutoff, even when walking around
(weapon bobbing means that additional margin is needed). Because of
this change, the place where the fist "lands" is now offset to the
right, possibly slightly too much, but not so much that I think
it's a huge problem.

This fixes #95.