dehacked: Split long string replacements.
BEX allows long string replacements to be continued onto multiple
lines by appending a \ to the end of the line to indicate that the
string continues on the next line. Make use of this syntax to split
some of the really long string replacements and make the DEHACKED
lumps more readable.