dist: Add a startup shell script and manpage
Pretty basic shell script for finding a port and launching it
automatically.  It's nothing overly fancy, but shell portability is
strived for, since it should be capable of launching no matter what
operating system is in use, it should not use anything more advanced
than is available in Bourn shell/POSIX (no ksh/bash extensions).