levels: Remove DM33, DM34.
As per discussion on #112, we seem to have reached a consensus that
FreeDM should (for the time being) only have 32 levels. So remove
DM33 and DM34 which were recently added.

In the future we can maybe revisit this decision, but for now we
want to concentrate on having 32 levels of the highest quality. Extra
levels means more maintenance burden.

This fixes #112, #108, #107.