lumps: Add MAPINFO for FreeDM.
In ZDoom-based ports, this removes the text intermission screens
that are intended to show story text (which is redundant for FreeDM).
It also makes the game loop back to MAP01 after MAP30 so that you can
keep playing endlessly.

Unfortunately this doesn't work in Odamex or ZDaemon, but it has been
tested in ZDoom and Zandronum. Odamex and ZDaemon are based on an
older version of ZDoom that does not support the newer MAPINFO
syntax, only an older, Hexen-style syntax. This old syntax does not
appear to be sufficient to accomplish what is desired here - at least,
not without potentially breaking things like Dehacked patches that
change the level name.

This fixes #111.