levels: replace Wolf-themed DM31/32 with nub_hat maps
As per #114, it is generally decided that FreeDM should not contain
remakes of other levels.  These weren't remakes, but definitely
Wolfenstein 3D homages, which don't really make interesting levels.
nub_hat had previously submitted these two maps for DM33/34 but they
were swiftly removed when it was decided to reduce FreeDM to a limited
number of maps (32), so let's replace the Wolf levels with them.