README: revise "Submitting your work"
Git is not within the reach of most contributors, so let's
de-emphasize its importance.  It would be nice if more people
submitted things via pull requests, but it's a fact-of-life for
Freedoom committing to be done by more technically-minded people
(typically its maintainers).  Point instead to the forums, issue
tracker, and file sharing sites as the common and accepted means of
submitting work.