README: rewrite/revisement of many sections
The introduction uses wording mostly copied from the web site to
maintain consistency within the project, and the usage of the term
"IWAD" is greatly reduced.  I still define it, but only so that
someone might become familiar with the term when other people use it,
rather than as a primary term.  Like wise, "port" and "source port"
are largely replaced with "engine," but while defining the former so a
newbie can understand what community members are speaking of.

"How to play" has some additional information about how to actually
get an engine to play the game.  It was previously lacking.  Odamex is
the only port actually outright named as a recommendation, but I try
to keep from sounding like it's the _only_ choice.

The "general rules" to follow has been reduced from four to three, the
first one heavily rewriting so as to not imply that derived works are
completely forbidden.

The "Using Git" section also had a major reworking.  I don't really
like the section entirely myself.  Either needs more revisement or
ejected entirely.

Quite a few linguistic tweaks in all of the sections.

Game names are more consistently italicized.