sounds/dspdiehi: Use the Wilhelm scream
Pretty famous movie sound effect, kind of hilarious to hear it.  Using
it on one of the less common sound effects to be heard -- the player
needs to die obtaining -50% health or less while not being gibbed.

The sound effect is in the public domain:
https://archive.org/details/WilhelmScreamSample

Some background info on the effect:
https://en.wikipedia.org/wiki/Wilhelm_scream