C1M1: Removed "collapsing wall"
The "collapsing wall" effect in the first hallway in the game was poorly
done. It just fell instantly with no sound effects or animation, and
just ended up looking like a visual anomaly. Not something you want to
show the player on their first impression, so I took it out.