patches: eliminate symlink for W108_1.
This texture is supposed to have a monster corpse hanging on the wall,
but is currently just an empty marble stone wall, and so is a symlink to
avoid duplication. But for simplicity it's probably easier to just
include a copy of the file; this will make it easier to be improved in
the future.