sounds: Remove dummy-beep.wav.
This is an old relic of the project: long ago when the project had
a shortage of sound effects, any missing sound effects would be
substituted with this recording I made of the PC speaker 'beep' from
my Linux desktop. The idea was that the beeping would alert people
that the sound effect was missing and encourage it to be replaced.

This idea ultimately proved counterproductive as the beeping just
proved annoying for anyone wanting to try to play the game. I later
replaced it with a dummy.wav that is just silent.