scripts: Remove sanity-check script.
This script was previously used for checking the correctness of the
assignment list files used for coordination of work in the earlier
stages of the project. As these files were removed over a year ago
in commit 7df4a6881c114400eaf3ef6cd8428d47a5b620d2, this script is
now obsolete.