textures: Add Doom 1 versions of conflicting textures.
Some textures which appear in both doom.wad and doom2.wad are defined
differently in each. We were always using the same definition in both
Phase 1 and Phase 2, causing compatibility problems with some levels
which expected a particular version of the texture.

To resolve, add both versions of the conflicting textures, controlled by
a new DOOM1_VERSIONS #define that allows which the version to be
selected. Use the doom.wad versions for Freedoom: Phase 1, and the
doom2.wad versions for Freedoom: Phase 2. This resolves part of #194.