dist: Remove boom, zdoom, prboom.  Add crispy-doom.
‘boom’ and ‘doom’ represent Debian-style links for the alternatives
system to set up preferred ports.  Now that we no longer require
Boom compatibility, it should be safe to remove the former. This
should only be an issue if chocolate-doom -> doom, and tries to play
one of the maps that aren’t vanilla-compatible yet.

zdoom and prboom are dead ports, superceded by gzdoom and prboom-plus.
I hope their removals aren’t controversial.  If it adversally affects
anybody: setup the ~/.doomport symlink to point to them.

Finally, crispy-doom.  Odamex is still preferred before it because of
its simpler multiplayer capabilities, otherwise Crispy would have been
a winner (imo).  It has a close relationship to Chocolate Doom, but I’m
leaving that last because of the non-vanilla compatibility of Freedoom.
Chocolate Doom is a last effort to run something.