c1m7: fix unreachable spectre using lower floor to lowest
Revert to previous version then change type of line 655 from W1 lower to
highest adjacent floor, to W1 lower to lowest.

It's clear from examining the map that sector 247, the platform with the
two sergeants, is meant to lower as you hurl yourself across and grab
the red key in mid-flight, revealing the spectre in sector 246. However,
the linedef action incorrectly selected the _highest_ adjacent floor as
destination, selecting the height of sector 195, and not moving at all.

The previous fix for this took the sledgehammer approach and deleted the
whole thing, which I felt was a shame.