c1m7: add/restore secrets and bullets/shells choice
- Mark the "hidden in plain sight" backpack/plasma room, and heavy
armor, as secrets (sectors 595, 196); I found these not being secret,
when a tiny box with a single plasma charge is so, underwhelming. And as
I recall the heavy armor used to be secret in previous map versions.

- Restore the "ammo choice" (sectors 156/157) which are supposed to both
rise up when you step onto one, leaving the other inaccessible. Now it
rises twice as far and can kill you, which is fun! (and avoids need to
add a control sector and rebuild nodes, which would bloat the diff!)
Choose bullets; there are enough shells on this map that missing a box
doesn't matter - although if you are fast enough you can still get both.