Deemphasize Final Doom compatibility.
Given that there are a handful of animations that are fundamentally
incompatible between Doom II and Final Doom, and we have decided that
the correct thing to do is to prefer Doom II compatibility.
Additionally, there are only maybe a half dozen Final Doom mods around
and they pretty much all focus on being unofficial sequels.  We won’t
remove nor continue to improve the existing TNT/Plutonia-specific
resources, let’s just make it a little less of an issue when those
mods are broken (assuming it’s due to the Doom II conflicts...).