graphics: tweaks on freedoom guy standard HUD face
Doom engines stretch the the images with 20% extra height. This,
together with the already slim jaw and sharp chin in the face of the
freedoom guy, makes it look very juvenile, stretched.

This change tweaks the face images to have a thicker jaw and a more mature
look. It also tweaks the lip position a bit and reduces the excessively
dark outline along the chin.