Complete Doom 3 BFG Support, and avoid crashing PrBoom-Plus
DOOM2.WAD contains a MAP33, which is being represented here as only
the dummy map, and also M_EPI1 and M_EPI2 lumps for the “Hell on Earth”
and “No Rest for the Living” texts, both of which are now replicated
here as well (using “Phase 2” and “The Missing Levels” as names instead).

No, we still don’t plan on having a real MAP33, the inclusion here is
only to avoid engine crashes.

Fixes: #533
Ref: #529