Further MAP33 support
Adds a test map to take the MAP33 slot, which merges MAP01 and MAP02 of test_levels.wad. Extra stuff has been added to the map. DeHackEd file was changed to reflect the change. ZMAPINFO file was added for Phase 2 to support MAP33. FreeDM's ZMAPINFO file was edited to do the same.