Chapter 4 Update
Updated Chapter 4: Double Impact with the fixes included in the
hotfix. C4M7 was already updated for Freedoom with the fixes, but not
C4M4 and C4M8. Both were updated to be up to par with the
hotfix. C4M8's old edits to lowering the lifts won't be needed
anymore. C4M8 was also edited to be vanilla-compatible, removing some
details at one of the final rooms. Hopefully Vanilla won't crash
anymore.

The endpic for Chapter 4 has been updated to match with the current
Freedoom assets, bye bye that old Imp sprite and floor texture.