dehacked: Turn on full brightness for a few frames
The firing frames for the zombie and assault tripod look weird if they're not 
shown in full brightness, so add a dehacked tweak to set them to full
brightness. Doom has the same bug but it isn't quite as noticeable.

This could theoretically cause a mod compatibility issue since unexpected
frames might be fully lit. However, (1) the worst that can happen is a minor
cosmetic issue, and (2) I don't know of any actual mods this affects (maybe
there aren't any). I think the benefit outweighs the potential downside.