sprites: improved player front walk animation
Reworked the legs so it looks more like actual walking.

I also reduced a bit the sudden swing with the weapon that
it does in the middle of the walk.

I've made  it so it actually looks like he's holding the gun
rather than just punching it. And I've tweaked a bit the helmet
and some minor details.