sprites: improve player back walk animation
The legs were reworked as well, matching the changes done to the front.

The shoulder pads movement was toned down, and the left arm was changed
so it's movement is a bit less abrupt and can actually be followed.
Other minor tweaks and size was made proportional to the front frames.