freedm: Remove override for menu skull.
FreeDM had a special override to use an animated rocket for the menu
selector, a tweak to make it more deathmatch-appropriate. Unfortunately
the rocket animation is too wide and causes Chocolate Doom to error out
when the help screen is brought up (F1) with "Bad V_DrawPatch". Even if
the horizontal offset is adjusted it will overlap with parts of the help
screen and obscure them. I think the easiest solution is to just remove
the override.