More intermission text adjustments
I still hate that "...", but I'm starting to think the aggravation is not really that itself but the use of a direct quote for "you" the player for something they never actually said.

So looking at the E4 text again to see how it could be avoided, I noticed a bunch of other little weird things... the past tense on the first sentence clashing with the rest of the present tense; the "despite" doesn't really fit with the logical progression; and while I know exactly what "plump down" was intended to convey it seems a bit nonstandard. It's also missing punctuation at the end, in contrast with the other intermission texts.

Here's a somewhat significant rewrite to get around those issues.

I've also made smaller changes to C5 and C6 for similar reasons.