documentation: change ascii ' quotes to either ’ or _
Non-reference AsciiDoc implementations such as AsciiDoctor (used by
GitHub) or even the simplistic parser in loccount choke up on these
characters for different reasons.  AsciiDoctor doesn’t seem to even
parse them, displaying them as-is and not with emphasized text.
loccount would get hung up on possessives, expecting a terminating
character for emphasis that never comes.

Annoying, but easy to work around.  the curly ’ character can be used
for possessives and _underscores_ can be used for emphasis.