Remove Doom 3: BFG Edition support
Might have been a nice idea to be able to run under absolutely any
Doom engine, this has caused problems with PrBoom-Plus for
compatibility, with that engine detecting our game as a BFG Edition
IWAD and applying certain workarounds that don’t actually make sense
in this case.

It is a rather niche use case to run under BFG Edition, and proper
PrBoom-Plus compatibility is much preferred.