test_levels: base E1M1 off of MAP01
Makes a little bit more of a consistent testing experience.  As of
this commit, there are a lot of HOMs/missing textures thanks to
depending on stuff presently in Phase 2 but not 1, though we’ll be
changing that too.

Removed MAP02, which was Cato’s old testing map and unnecessary to
stick around.