textures: include almost all textures in all IWADs
We have been operating since the beginning with the idea of only
including matching compatible textures for Phase 1 and 2 based on the
textures that appeared in Doom 1 and 2, including keeping some
exclusive to each game.

This is artifically limiting to map creators and there is no good
reason to keep it this way.  Fixes #588