dist/freedoom: support ~/.local installs, make it easier to read paths
Bunch of logic to use XDG_DATA_HOME or fallback according to the
actual XDG recommendations.

Use an array to build up paths (can be edited easily in the future or
for whatever custom purpose), which shellcheck warned is a bash-ism,
so change the script interpreter to bash as well.  Let’s hope that
doesn’t break anyone (does anyone lack bash even if they don’t use it
as their shell?)