textures: redefine A-BROWN4 so it doesn’t use BOSFA0
With WAD merging with certain Doom 2 mods (eg,
doom2/Ports/megawads/strg), using a sprite in this texture causes some
engines to crash on loading the game.  The mod makes up its new BOSF*
sprites but omits BOSFA0, which caused our game to crash when trying
to load Freedoom with Struggle.

This can break Plutonia mod compatibility a bit if they try to replace
this sprite too, but let’s hope that will not happen.