make: reduce install targets to just “freedm” and “freedoom”
Causes the installation to be more similar to how the pre-built Zip
files are handled: separating out the multiplayer and
singleplayer-focused games into two targets.  This is similar also to
how Debian and Fedora generate their packages for the game and it is a
sensible one.

Also creates a lot of duplicated lines in the make rules, even though
the number of targets is actually reduced.