map11: Fix unreachable SKAG and monsters
Removed unreachable SKAG near the start as well as the bars in front
of it. Added a nominal health potion for symmetry with the pillar to
the right without overly affecting balance.

Near the end of the map there are six monsters on pillars that can get
knocked off. Once knocked off, either from player or monster
projectiles, they're very hard to kill making 100% kills difficult.
Teleport them to where the player can kill them when they get knocked
off.