New map by Mortrixs slot 21.
Mortrixs: "the map is a replacement for Map21, because the current one is bad"

I (xindage) played the map and I find no issues on coop/singleplay, dm spots are present so the map is fine in my point of view.
Mortrixs told me too to keep the original slot name for the map called "Under realm".