Give MAP31 normal and secret exits
This changes the old exit to be the normal exit, and the old fake exit
to be the proper secret exit. Access to the secret exit requires
getting all the secrets in the map (i.e. telefragging all the Assault
Tripods/Cyberdemons).

I also fixed most of the switches not visually switching in vanilla.