Extra edits on the following maps: (Not all of this is needed to be documented in the main changelog.)
dm01 - Corrected the alignment in the exit room.
dm04 – Removed some unused linedefs.
dm07 - Corrected the alignment in the exit room.
dm09 - Corrected the alignment in the exit room.
dm11 - Reworked most of the south side of map and added an invisibility sphere.
dm14 - Retextured the exit to match the theme and added a chainsaw there.
dm27 - Corrected the alignment in the exit room and added missing chainsaw.