New sprites for Summoner and Dead Flare
New monster sprites by Korp and HorrorMovieRei. New summoner which we dubbed "Matribite" due to it being based on the Trilobite, and a new Dead Flare replacement named "hatchling" since it is some sort of larvae.

Included is also a tinkered version of the death sound made by mc776 so it synchs better with the new death animation for the Matribite.