Change manual to build as A5 paper size.
This seems like a better choice for the manual since it is a "booklet"
size similar to the original Doom manual, rather than a "full page"
size that is rather unusual for a book.

Several changes are needed to allow asciidoctor to successfully produce
a PDF. The main change is to set a bunch of images to 100% width,
especially in tables where this means "100% of cell width". This allows
the tables to scale down successfully to the smaller size. I have also
adjusted the column widths of various tables to accommodate the smaller
size without the text wrapping awkwardly.