Change skull key colorblind patterns to match keycards.
The idea here is that the horns on the skulls indicate the same shapes
as those used for the keycards:

* Blue: X shape
* Red: Horizontal (horns point horizontally atop head)
* Yellow: Vertical

In order to do this I color-swapped and exchanged the graphics for the
red and blue skull keys. This commit updates the skull key sprites,
status bar key icons and the skull key wall textures. As well as as the
color swap, I've also tweaked the skulls to increase emphasis on the
horn directions.